pallet_subspace_mmr
Trait Config

pub trait Config: Config<Hash: Into<H256> + From<H256>> {
    type MmrRootHash: Parameter + Copy + MaxEncodedLen;
    type MmrRootHashCount: Get<u32>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type MmrRootHash: Parameter + Copy + MaxEncodedLen

type MmrRootHashCount: Get<u32>

The number of mmr root hashes to store in the runtime. It will be used to verify mmr proof statelessly and the number of roots stored here represents the number of blocks for which the mmr proof is valid since it is generated. After that the mmr proof will be expired and the prover needs to re-generate the proof.

Object Safety

This trait is not object safe.